+++ app/Lang.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE QuasiQuotes #-}
@@ -9,13 +10,16 @@ module Lang where
 import Data.Monoid (First)
 import Language.Nanopass.QQ (deflang)
 
+data Foo a b c = Foo [c]
+  deriving (Show,Functor,Foldable,Traversable)
+
 [deflang| L0 (funny)
   (Expr
     (Var {x String})
     (Lam {x String} {e ($Stmt *)})
     (App {f $Expr} {a $Expr})
     (Nope String)
-    (UhOh {(First $Expr *) ($Expr *)})
+    (UhOh {(First $Expr *) ($Expr *) (Foo Int Int $Expr)})
   )
   (Stmt
     (Expr {delme funny} $Expr)